feat: add custom hooks and styled components for improved functionality and UI, including useDebouncedValue, useService, and pagination components
