+++ src/pages/home/home.tsx
@@ -1,42 +1,14 @@
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { getMovies, searchMovies } from "@movie/services/api/movie-api";
-import { useService } from "@app/lib/useService";
+import { useService } from "@app/lib/use-service";
 import { Card } from "@app/lib/components/card/card";
-import { Search } from "@app/lib/components/search/search";
 import { Pagination } from "@app/lib/components/pagination/pagination";
 import { useState } from 'react';
 import { MovieCardsSkeleton } from '@app/lib/components/skeleton-cards-list/cards-skeleton-list';
 import { ErrorComponent } from '@core/error/components/error-component';
-import styled from 'styled-components';
 import { scrollToTop } from '@app/utils/scroll-top';
-
-const ResetButton = styled.button`
-  position: fixed;
-  top: 1rem;
-  right: 1rem;
-  padding: 0.75rem 1.5rem;
-  background: #007bff;
-  border: none;
-  border-radius: 8px;
-  color: white;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  font-weight: 500;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  z-index: 100;
-
-  &:hover {
-    background: #0056b3;
-    transform: translateY(-1px);
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  }
-
-  &:active {
-    transform: translateY(0);
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  }
-`;
-
+import { NavSection } from './components/nav-section';
+import { ContentWrapper, StickyContainer, MoviesContainer, MoviesGrid } from './home.styled';
 
 export const Home: React.FC = () => {
   const navigate = useNavigate();
@@ -68,62 +40,53 @@ export const Home: React.FC = () => {
     scrollToTop();
   };
 
-  const handleReset = () => {
-    setSearchParams({});
-    scrollToTop();
-  };
-
   return (
     <>
-      <ResetButton onClick={handleReset}>
-        Home
-      </ResetButton>
-      <Search onSearch={setSearchQuery} />
-      {error ? (
-        <ErrorComponent code={error.code} message={error.message} />
-      ) : loading ? (
-        <MovieCardsSkeleton />
-      ) : !movies?.length ? (
-        <h3>No movies found</h3>
-      ) : (
-        <>
-          <div style={{ padding: '2rem' }}>
-            <div style={{ 
-              display: 'grid',
-              gridTemplateColumns: 'repeat(auto-fill, minmax(180px, 1fr))',
-              gap: '2rem',
-              placeItems: 'center'
-            }}>
-              {movies.map((movie) => (
-                <Card 
-                  key={movie.id}
-                  onClick={() => handleMovieClick(movie.id)}
-                >
-                  <Card.Image 
-                     src={`https://image.tmdb.org/t/p/w220_and_h330_face${movie.posterPath}`}
-                     alt={movie.title}
-                     loading="lazy"
-                     score={Math.floor(movie.voteAverage * 10)}
-                  />
-                  <Card.Body>
-                    <Card.Title>{movie.title}</Card.Title>
-                  </Card.Body>
-                  <Card.Footer>
-                    <small>Release Date: {movie.releaseDate}</small>
-                  </Card.Footer>
-                </Card>
-              ))}
-            </div>
-          </div>
-          {totalPages && totalPages > 1 && (
-            <Pagination
-              currentPage={currentPage}
-              totalPages={totalPages}
-              onPageChange={handlePageChange}
-            />
-          )}
-        </>
-      )}
+      <StickyContainer>
+        <NavSection setSearchQuery={setSearchQuery} />
+      </StickyContainer>
+      <ContentWrapper>
+        {error ? (
+          <ErrorComponent code={error.code} message={error.message} />
+        ) : loading ? (
+          <MovieCardsSkeleton />
+        ) : !movies?.length ? (
+          <h3>No movies found</h3>
+        ) : (
+          <>
+            <MoviesContainer>
+              <MoviesGrid>
+                {movies.map((movie) => (
+                  <Card 
+                    key={movie.id}
+                    onClick={() => handleMovieClick(movie.id)}
+                  >
+                    <Card.Image 
+                       src={`https://image.tmdb.org/t/p/w220_and_h330_face${movie.posterPath}`}
+                       alt={movie.title}
+                       loading="lazy"
+                       score={Math.floor(movie.voteAverage * 10)}
+                    />
+                    <Card.Body>
+                      <Card.Title>{movie.title}</Card.Title>
+                    </Card.Body>
+                    <Card.Footer>
+                      <small>Release Date: {movie.releaseDate}</small>
+                    </Card.Footer>
+                  </Card>
+                ))}
+              </MoviesGrid>
+            </MoviesContainer>
+            {totalPages && totalPages > 1 && (
+              <Pagination
+                currentPage={currentPage}
+                totalPages={totalPages}
+                onPageChange={handlePageChange}
+              />
+            )}
+          </>
+        )}
+      </ContentWrapper>
     </>
   );
 };
